feat: remove distinctions between KakarotZero and KakarotSSJ
