+++ crates/ef-testing/src/evm_sequencer/sequencer/mod.rs
@@ -1,22 +1,4 @@
-#[cfg(feature = "v0")]
-pub mod v0;
-#[cfg(feature = "v1")]
-pub mod v1;
-
 use blockifier::bouncer::BouncerConfig;
-#[cfg(feature = "v0")]
-pub use v0::INITIAL_SEQUENCER_STATE;
-
-#[cfg(feature = "v1")]
-pub use v1::INITIAL_SEQUENCER_STATE;
-
-#[cfg(not(any(feature = "v0", feature = "v1")))]
-use lazy_static::lazy_static;
-#[cfg(not(any(feature = "v0", feature = "v1")))]
-lazy_static! {
-    pub static ref INITIAL_SEQUENCER_STATE: State = State::default();
-}
-
 use starknet::core::types::Felt;
 use std::ops::{Deref, DerefMut};
 
@@ -39,12 +21,33 @@ use starknet_api::{
     core::{ChainId, ClassHash, ContractAddress},
 };
 use std::num::NonZeroU128;
-
-use super::{
-    constants::{ETH_FEE_TOKEN_ADDRESS, STRK_FEE_TOKEN_ADDRESS},
+use crate::evm_sequencer::{
+    constants::{
+        storage_variables::{
+            ACCOUNT_PUBLIC_KEY, ERC20_BALANCES, KAKAROT_ACCOUNT_CONTRACT_CLASS_HASH,
+            KAKAROT_BLOCK_GAS_LIMIT,
+            KAKAROT_NATIVE_TOKEN_ADDRESS, KAKAROT_UNINITIALIZED_ACCOUNT_CLASS_HASH, OWNABLE_OWNER,
+        },
+        ACCOUNT_CONTRACT_CLASS, ACCOUNT_CONTRACT_CLASS_HASH, BLOCK_GAS_LIMIT, ETH_FEE_TOKEN_ADDRESS, FEE_TOKEN_CLASS, FEE_TOKEN_CLASS_HASH,
+        KAKAROT_ADDRESS, KAKAROT_CLASS, KAKAROT_CLASS_HASH, KAKAROT_OWNER_ADDRESS,
+        OPENZEPPELIN_ACCOUNT_CLASS, OPENZEPPELIN_ACCOUNT_CLASS_HASH, RELAYER_ADDRESS,
+        RELAYER_BALANCE, RELAYER_VERIFYING_KEY, UNINITIALIZED_ACCOUNT_CLASS,
+        UNINITIALIZED_ACCOUNT_CLASS_HASH, STRK_FEE_TOKEN_ADDRESS,
+    },
     types::contract_class::CasmContractClassWrapper,
     utils::compute_starknet_address,
 };
+
+#[cfg(feature = "v0")]
+use crate::evm_sequencer::constants::{CAIRO1_HELPERS_CLASS, CAIRO1_HELPERS_CLASS_HASH, storage_variables::KAKAROT_CAIRO1_HELPERS_CLASS_HASH};
+
+use blockifier::abi::abi_utils::get_storage_var_address;
+#[allow(unused_imports)]
+use blockifier::state::state_api::{
+    State as BlockifierState, StateReader as BlockifierStateReader,
+};
+use lazy_static::lazy_static;
+use sequencer::state::State as SequencerState;
 
 /// Kakarot wrapper around a sequencer.
 #[derive(Clone)]
@@ -196,4 +199,88 @@ pub fn convert_contract_class_v1(class: &CompiledClass) -> Result<ContractClass,
     Ok(ContractClass::V1(ContractClassV1::try_from(
         casm_contract_class,
     )?))
+}
+
+lazy_static! {
+    pub static ref INITIAL_SEQUENCER_STATE: SequencerState = {
+        let mut state = SequencerState::default();
+
+        // On KakarotZero we rely on the Cairo1 helpers class for unavailable syscalls and precompiles.
+        #[cfg(feature = "v0")]
+        let storage = vec![
+            (OWNABLE_OWNER, *KAKAROT_OWNER_ADDRESS.0.key()),
+            (KAKAROT_NATIVE_TOKEN_ADDRESS, *ETH_FEE_TOKEN_ADDRESS.0.key()),
+            (KAKAROT_ACCOUNT_CONTRACT_CLASS_HASH, ACCOUNT_CONTRACT_CLASS_HASH.0),
+            (KAKAROT_CAIRO1_HELPERS_CLASS_HASH, CAIRO1_HELPERS_CLASS_HASH.0),
+            (KAKAROT_BLOCK_GAS_LIMIT, Felt::from(BLOCK_GAS_LIMIT)),
+            (KAKAROT_UNINITIALIZED_ACCOUNT_CLASS_HASH, UNINITIALIZED_ACCOUNT_CLASS_HASH.0),
+        ];
+
+        #[cfg(not(feature = "v0"))]
+        let storage = vec![
+            (OWNABLE_OWNER, *KAKAROT_OWNER_ADDRESS.0.key()),
+            (KAKAROT_NATIVE_TOKEN_ADDRESS, *ETH_FEE_TOKEN_ADDRESS.0.key()),
+            (KAKAROT_ACCOUNT_CONTRACT_CLASS_HASH, ACCOUNT_CONTRACT_CLASS_HASH.0),
+            (KAKAROT_BLOCK_GAS_LIMIT, Felt::from(BLOCK_GAS_LIMIT)),
+            (KAKAROT_UNINITIALIZED_ACCOUNT_CLASS_HASH, UNINITIALIZED_ACCOUNT_CLASS_HASH.0),
+        ];
+
+        // Write all the storage vars to the sequencer state.
+        for (k, v) in storage {
+            (&mut state).set_storage_at(*KAKAROT_ADDRESS, get_storage_var_address(k, &[]), v).expect("failed to set storage");
+        }
+
+        // Convert the contract classes to using the util corresponding to the Cairo version it's compiled with.
+        let (converted_kakarot_class, converted_account_class, converted_uninitialized_class) = {
+            #[cfg(not(feature = "v1"))]
+            {
+                // Default to Cairo 0
+                (
+                    convert_contract_class_v0(&KAKAROT_CLASS).expect("failed to convert kakarot class"),
+                    convert_contract_class_v0(&ACCOUNT_CONTRACT_CLASS).expect("failed to convert account class"),
+                    convert_contract_class_v0(&UNINITIALIZED_ACCOUNT_CLASS).expect("failed to convert uninitialized class")
+                )
+            }
+            #[cfg(feature = "v1")]
+            {
+                (
+                    convert_contract_class_v1(&KAKAROT_CLASS).expect("failed to convert kakarot class"),
+                    convert_contract_class_v1(&ACCOUNT_CONTRACT_CLASS).expect("failed to convert account class"),
+                    convert_contract_class_v1(&UNINITIALIZED_ACCOUNT_CLASS).expect("failed to convert uninitialized class")
+                )
+            }
+        };
+
+        // Write the kakarot class and class hash.
+        (&mut state).set_class_hash_at(*KAKAROT_ADDRESS, *KAKAROT_CLASS_HASH).expect("failed to set kakarot class hash");
+        (&mut state)
+            .set_contract_class(*KAKAROT_CLASS_HASH, converted_kakarot_class).expect("failed to set kakarot contract class");
+
+        // Write contract account, uninitialized_account and erc20 classes and class hashes.
+        (&mut state).set_contract_class(
+            *ACCOUNT_CONTRACT_CLASS_HASH,
+            converted_account_class,
+        ).expect("failed to set contract account class");
+        (&mut state)
+            .set_contract_class(*UNINITIALIZED_ACCOUNT_CLASS_HASH, converted_uninitialized_class).expect("failed to set eoa contract class");
+
+        (&mut state).set_contract_class(
+            *FEE_TOKEN_CLASS_HASH,
+            convert_contract_class_v0(&FEE_TOKEN_CLASS).expect("failed to convert FEE TOKEN CLASS to contract class"),
+        ).expect("failed to set sequencer contract class");
+        (&mut state).set_class_hash_at(*ETH_FEE_TOKEN_ADDRESS, *FEE_TOKEN_CLASS_HASH).expect("failed to set fee token class hash");
+
+        #[cfg(feature = "v0")]
+        (&mut state).set_contract_class(*CAIRO1_HELPERS_CLASS_HASH, convert_contract_class_v1(&CAIRO1_HELPERS_CLASS).expect("failed to convert CAIRO1_HELPERS Class to contract class")).expect("failed to set cairo1_helpers contract class");
+
+        (&mut state).set_contract_class(
+            *OPENZEPPELIN_ACCOUNT_CLASS_HASH,
+            convert_contract_class_v0(&OPENZEPPELIN_ACCOUNT_CLASS).expect("failed to convert OPENZEPPELIN ACCOUNT CLASS to contract class")
+        ).expect("failed to set openzeppelin account contract class");
+        (&mut state).set_class_hash_at(*RELAYER_ADDRESS, *OPENZEPPELIN_ACCOUNT_CLASS_HASH).expect("failed to set relayer class hash");
+        (&mut state).set_storage_at(*RELAYER_ADDRESS, get_storage_var_address(ACCOUNT_PUBLIC_KEY, &[]), RELAYER_VERIFYING_KEY.scalar()).expect("failed to set relayer public key");
+        (&mut state).set_storage_at(*ETH_FEE_TOKEN_ADDRESS, get_storage_var_address(ERC20_BALANCES, &[*RELAYER_ADDRESS.0.key()]), *RELAYER_BALANCE).expect("failed to set relayer balance");
+
+        state
+    };
 }
